Renamed mmap.py to files.py, moved TempAnalysisFile to files.py and renamed mmapFile to LargeFileReader
