+++ services.py
@@ -1,6 +1,4 @@
 import os
-import tempfile
-import shutil
 import sys
 
 # correctly import renamed modules (Py2 vs Py3)
@@ -75,37 +73,6 @@ class ServiceMeta():
     def __iter__ (self):
         for key in self.data:
             yield (key, self.data[key])
-
-
-
-class TempAnalysisFile(object):
-    """
-    Temporary Analysis File class.
-    """
-
-    def __init__(self, obj):
-        self.obj = obj
-        #print(obj)
-
-    def __enter__(self):
-        """
-        Create the temporary file on disk.
-        """
-
-        tempdir = tempfile.mkdtemp()
-        self.directory = tempdir
-        tfile = os.path.join(tempdir, self.obj)
-        with open(tfile, "wb") as f:
-            f.write(open(self.obj).read().encode())
-        return tfile
-
-    def __exit__(self, type, value, traceback):
-        """
-        Cleanup temporary file on disk.
-        """
-
-        if os.path.isdir(self.directory):
-            shutil.rmtree(self.directory)
